refactor(repositories): hoist assert_no_residual_module_syntax to prompt_artifact
Closes #198.

\`assert_no_residual_module_syntax\` was identically duplicated in
\`skill.rs\` and \`project_prompt.rs\`. Both bodies walked the same
\`code_fence_pattern\` / \`residual_patterns\` regex setup and built
the same error message — any bug fix had to be applied in two
places.

Hoist the function into \`prompt_artifact.rs\` (which both consumers
already import from for \`read_prompt_artifact\`) and re-export it.
Drop both copies; \`use ::{assert_no_residual_module_syntax,
read_prompt_artifact}\` covers the call sites.

Test coverage moves with the helper — five new tests in
\`prompt_artifact::tests\` lock in the contract:

  - clean markdown passes
  - \`export default\` is rejected with file:line in the message
  - \`import\` is rejected with file:line
  - fenced code (\`\`\`js / \`\`\`) examples are skipped (legitimately
    quoted JS, not residual module syntax)
  - tilde fences (\`~~~ts\`) are also skipped

`cargo test --lib repositories` is green at 50/50 (45 prior + 5
new prompt_artifact tests).
