+++ sdk/src/repositories/prompt_artifact.rs
@@ -105,9 +105,53 @@ pub fn read_prompt_artifact(
   }
 }
 
+/// Walk a compiled prompt's text and reject any leftover ESM-style
+/// module syntax (`export default`, `export const`, `import`) that
+/// should have been stripped by `mdx_to_md_with_metadata`. Lines
+/// inside fenced code blocks (``` / ~~~) are skipped so legitimately
+/// quoted JS examples don't trip the check.
+///
+/// Pre-#198 the same body was duplicated in both `skill.rs` and
+/// `project_prompt.rs`; bug fixes had to be applied in lockstep.
+/// Centralising here makes the assertion the single source of truth.
+pub fn assert_no_residual_module_syntax(content: &str, file_path: &str) -> Result<(), String> {
+  let code_fence_pattern = regex_lite::Regex::new(r"^\s*(```|~~~)").unwrap();
+  let residual_patterns = [
+    regex_lite::Regex::new(r"^\s*export\s+default\b").unwrap(),
+    regex_lite::Regex::new(r"^\s*export\s+const\b").unwrap(),
+    regex_lite::Regex::new(r"^\s*import\b").unwrap(),
+  ];
+  let mut active_fence: Option<&str> = None;
+  for (index, line) in content.lines().enumerate() {
+    if let Some(caps) = code_fence_pattern.captures(line) {
+      let marker = caps.get(1).map(|m| m.as_str()).unwrap_or("");
+      if active_fence.is_none() {
+        active_fence = Some(marker);
+      } else if active_fence == Some(marker) {
+        active_fence = None;
+      }
+      continue;
+    }
+    if active_fence.is_some() {
+      continue;
+    }
+    for pat in &residual_patterns {
+      if pat.is_match(line) {
+        return Err(format!(
+          "Compiled prompt still contains residual module syntax at {}:{}: {}",
+          file_path,
+          index + 1,
+          line.trim()
+        ));
+      }
+    }
+  }
+  Ok(())
+}
+
 #[cfg(test)]
 mod tests {
-  use super::should_compile_dist_artifact;
+  use super::{assert_no_residual_module_syntax, should_compile_dist_artifact};
 
   #[test]
   fn plain_markdown_dist_prompt_skips_compilation() {
@@ -123,5 +167,40 @@ mod tests {
     ));
     assert!(should_compile_dist_artifact("Platform: {os.platform}\n"));
     assert!(should_compile_dist_artifact("<Md when={true}>ok</Md>\n"));
+  }
+
+  #[test]
+  fn assert_no_residual_module_syntax_passes_clean_markdown() {
+    let body = "# Title\n\nA paragraph.\n";
+    assert!(assert_no_residual_module_syntax(body, "ok.md").is_ok());
+  }
+
+  #[test]
+  fn assert_no_residual_module_syntax_rejects_export_default() {
+    let body = "# Title\nexport default { scope: 'project' }\n";
+    let err = assert_no_residual_module_syntax(body, "bad.md").unwrap_err();
+    assert!(err.contains("residual module syntax"));
+    assert!(err.contains("bad.md:2"));
+  }
+
+  #[test]
+  fn assert_no_residual_module_syntax_rejects_import() {
+    let body = "import x from 'y'\n# Title\n";
+    let err = assert_no_residual_module_syntax(body, "bad.md").unwrap_err();
+    assert!(err.contains("bad.md:1"));
+  }
+
+  #[test]
+  fn assert_no_residual_module_syntax_skips_code_fences() {
+    // `import …` inside a fenced JS example is documentation, not
+    // residual module syntax.
+    let body = "# Title\n\n```js\nimport { foo } from 'bar';\n```\n";
+    assert!(assert_no_residual_module_syntax(body, "ok.md").is_ok());
+  }
+
+  #[test]
+  fn assert_no_residual_module_syntax_skips_tilde_fences() {
+    let body = "# Title\n\n~~~ts\nexport default {}\n~~~\n";
+    assert!(assert_no_residual_module_syntax(body, "ok.md").is_ok());
   }
 }
